throw UnexpectedResponseException
This is a proposed breaking change where the conversation engine can throw a new exception type UnexpectedResponseException in addition to UnmatchedResponseException. The idea is to be able to differentiate between a response to a question that was not understood and a user starting a conversation when the bot is not expecting a response.